requirements.txt
# Core Libraries
azureml-core==1.42.0
azureml-sdk==1.42.0
azureml-train-core==1.42.0
azureml-train-automl-client==1.42.0
azureml-contrib-services==1.42.0

# Machine Learning Libraries
scikit-learn==1.0.2
lightgbm==3.3.2
xgboost==1.7.4
pandas==1.5.0
numpy==1.23.4

# API and Deployment Libraries
flask==2.2.2
requests==2.28.1

# Data Visualization Libraries
matplotlib==3.6.0
seaborn==0.12.0

# Azure Cosmos DB Integration
azure-cosmos==4.3.0

# DeepSeek-R1 (Hypothetical Library)
# Replace with the actual DeepSeek-R1 library if available
# deepseek-r1==1.0.0

# Other Utilities
python-dotenv==0.21.0
jupyter==1.0.0